[filescanner/library] Move filescanner to subfolder
