+++ src/Makefile.am
@@ -2,7 +2,7 @@
 sbin_PROGRAMS = forked-daapd
 
 if COND_ITUNES
-ITUNES_SRC=filescanner_itunes.c
+ITUNES_SRC=library/filescanner_itunes.c
 endif
 
 if COND_SPOTIFY
@@ -79,9 +79,9 @@ forked_daapd_SOURCES = main.c \
 	logger.c logger.h \
 	conffile.c conffile.h \
 	cache.c cache.h \
-	filescanner.c filescanner.h \
-	filescanner_ffmpeg.c filescanner_playlist.c \
-	filescanner_smartpl.c $(ITUNES_SRC) \
+	library/filescanner.c library/filescanner.h \
+	library/filescanner_ffmpeg.c library/filescanner_playlist.c \
+	library/filescanner_smartpl.c $(ITUNES_SRC) \
 	library.c library.h \
 	mdns_avahi.c mdns.h \
 	remote_pairing.c remote_pairing.h \
