+++ src/library/filescanner.h
